requirements.txt
Flask==2.3.2
pycups==2.0.1